python/ranker/requirements.txt
Cython>=2.0.0,<3.0
flask==3.1.2
git+https://github.com/twmht/python-rocksdb.git@v0.7.0
psycopg2-binary==2.9.11
numpy==2.3.5